// enums/table-chat-message-type.ts
export enum TableChatMessageType {
  CHAT = "CHAT",
  CIPHER_KEY = "CIPHER_KEY",
  F_KEY = "F_KEY",
  GAME_RATING = "GAME_RATING",
  HERO_CODE = "HERO_CODE",
  HERO_MESSAGE = "HERO_MESSAGE",
  TABLE_HOST = "TABLE_HOST",
  TABLE_TYPE = "TABLE_TYPE",
  USER_BOOTED = "USER_BOOTED",
  USER_GRANTED_SEAT_ACCESS_INVITEE = "USER_GRANTED_SEAT_ACCESS_INVITEE",
  USER_GRANTED_SEAT_ACCESS_INVITER = "USER_GRANTED_SEAT_ACCESS_INVITER",
  USER_JOINED = "USER_JOINED",
  USER_INVITED = "USER_INVITED",
  USER_LEFT = "USER_LEFT",
}